Refactor: #1 DNN 모델 전체적으로 리펙토링 진행
